requirements.txt
cmake==3.27.4.1
numpy==1.21.6
opencv-python==4.8.0.76
Pillow==9.5.0
face_recognition==1.3.0
face-recognition-models==0.3.0
Click==8.1.7
dlib>=19.7
